Move `sea-query-derive` into `sea-query-derive-internal`, making the internal logic can be reused
