+++ sea-query-derive-internal/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "sea-query-derive-internal"
+version = "0.3.0"
+authors = [ "Follpvosten <wolfi@karpador.xyz>", "Rene Leveille <rene@nestingsafe.com>" ]
+edition = "2021"
+description = "Derive macro for sea-query's Iden trait"
+license = "MIT OR Apache-2.0"
+documentation = "https://docs.rs/sea-query"
+repository = "https://github.com/SeaQL/sea-query"
+categories = [ "database" ]
+keywords = [ "database", "sql", "mysql", "postgres", "sqlite" ]
+rust-version = "1.60"
+
+[dependencies]
+syn = { version = "1", default-features = false, features = ["parsing", "proc-macro", "derive", "printing"] }
+quote = { version = "1", default-features = false }
+heck = { version = "0.4", default-features = false }
+proc-macro2 = { version = "1", default-features = false }
+thiserror = { version = "1.0", default-features = false }
+
+[dev-dependencies]
+trybuild = "1.0"
+sea-query = { version = "0", path = ".." }
+strum = { version = "0.24", features = ["derive"] }
